feat(dashboard): zoom risk trend chart
